CampusPulse/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
rq==1.15.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
videohash==3.0.1
pillow==10.1.0
numpy==1.24.4
scikit-learn==1.3.2
aiofiles==23.2.1
slowapi==0.1.9
python-dotenv==1.0.0